% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_predict.R
\name{multi_predict._mixo_spls}
\alias{multi_predict._mixo_spls}
\title{Model predictions across many sub-models}
\usage{
\method{multi_predict}{`_mixo_spls`}(object, new_data, type = NULL, num_comp = NULL, ...)
}
\arguments{
\item{num_comp}{An integer vector for the number of PLS terms to retain.}
}
\description{
Model predictions across many sub-models
}
