% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{multi_predict._mixo_pls}
\alias{multi_predict._mixo_pls}
\alias{multi_predict._mixo_spls}
\alias{multi_predict._mixo_plsda}
\alias{multi_predict._mixo_splsda}
\title{Model predictions across many sub-models}
\usage{
\method{multi_predict}{`_mixo_pls`}(object, new_data, num_comp = NULL, type = NULL, ...)

\method{multi_predict}{`_mixo_spls`}(object, new_data, num_comp = NULL, type = NULL, ...)

\method{multi_predict}{`_mixo_plsda`}(object, new_data, num_comp = NULL, type = NULL, ...)

\method{multi_predict}{`_mixo_splsda`}(object, new_data, num_comp = NULL, type = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{model_fit}}

\item{new_data}{A rectangular data object, such as a data frame.}

\item{num_comp}{An integer vector for the number of PLS terms to retain.}

\item{type}{A single character value or \code{NULL}. Possible values
are "numeric", "class", or "prob". When \code{NULL}, \code{predict()} will choose an
appropriate value based on the model's mode.}

\item{...}{Not currently used.}
}
\description{
Model predictions across many sub-models
}
