% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsmod-package.R
\docType{package}
\name{plsmod-package}
\alias{plsmod}
\alias{plsmod-package}
\title{plsmod: Model Wrappers for Projection Methods}
\description{
Bindings for additional regression models for use with the 'parsnip' package, including ordinary and spare partial least squares models for regression and classification (Rohart et al (2017) \doi{10.1371/journal.pcbi.1005752}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://plsmod.tidymodels.org}
  \item \url{https://github.com/tidymodels/plsmod}
  \item Report bugs at \url{https://github.com/tidymodels/plsmod/issues}
}

}
\author{
\strong{Maintainer}: Max Kuhn \email{max@posit.co} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
