% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_plsr.R
\name{plsr_fit}
\alias{plsr_fit}
\title{Wrapper for pls pls functions}
\usage{
plsr_fit(x, y, ncomp = NULL, ...)
}
\arguments{
\item{x}{A data frame or matrix of predictors.}

\item{y}{A matrix, vector, or data frame.}

\item{ncomp}{The number of PLS components. If left NULL, the maximum possible
is used.}
}
\value{
A model object generated by \code{\link[pls:mvr]{pls::plsr()}}.
}
\description{
Based on arguments, this wrapper preprocesses the data for \code{plsr} regression.
\code{plsr} takes a formula input, but requires x and y to be matrices in order to
handle multivariate predictors and outcomes. This wrapper thus takes
matrix \code{x} and \code{y} and converts them to the expected formula format
under the hood.
}
\keyword{internal}
