% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy.mixo_pls}
\alias{tidy.mixo_pls}
\alias{tidy.mixo_spls}
\title{Tidy methods for pls and spls objects}
\usage{
\method{tidy}{mixo_pls}(x, ...)

\method{tidy}{mixo_spls}(x, ...)
}
\arguments{
\item{x}{An object with class \code{mixo_pls} or \code{mixo_spls}.}

\item{...}{Not currently used.}
}
\value{
A tibble with columns \code{terms} (the predictor names), \code{value} (the
loadings), \code{type} (either "predictors" or "outcomes"), and \code{component} (the
component number).
}
\description{
Tidy methods for pls and spls objects
}
